requirements.txt
allure-pytest==2.13.2
pytest==7.3.1
selenium==4.9.0
webdriver-manager==3.8.6
